+++ lib/pico_phone/rails/railtie.rb
@@ -17,6 +17,13 @@ module PicoPhone
         end
       end
 
+      initializer "pico_phone_rails.active_job_serializer" do
+        ActiveSupport.on_load(:active_job) do
+          require "pico_phone/rails/serializers/phone_number_serializer"
+          ActiveJob::Serializers.add_serializers(PicoPhone::Rails::Serializers::PhoneNumberSerializer)
+        end
+      end
+
       initializer "pico_phone_rails.i18n" do |app|
         app.config.i18n.load_path += Dir[File.expand_path("locale/*.yml", __dir__)]
       end